big_data_analysis/assignment3/pyproject.toml
[tool.poetry]
name = "assignment3"
version = "0.1.0"
description = ""
authors = ["Aleksandr J. Smoliakov <aleksandr.smoliakov@mif.stud.vu.lt>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.12"
pymongo = "^4.12.1"
pandas = "^2.2.3"
matplotlib = "^3.10.1"
tqdm = "^4.67.1"
pyyaml = "^6.0.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
